refactor: rename CTA component
