+++ components/CTA/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as CTA } from "./CTA";
