+++ pages/[city]/[category]/[number]/index.tsx
@@ -1,7 +1,7 @@
 import cities, { ICategoryData } from "@mauriciorobayo/pyptron";
 import { TheMoneytizer } from "components/Ads";
 import { Aside } from "components/Aside";
-import CTA from "components/call-to-action";
+import { CTA } from "components/CTA";
 import { Header } from "components/Header";
 import { NumbersData } from "components/NumbersData";
 import Post from "components/post";
