+++ pages/[city]/[category]/index.tsx
@@ -1,8 +1,8 @@
 import cities from "@mauriciorobayo/pyptron";
 import { TheMoneytizer } from "components/Ads";
 import { Aside } from "components/Aside";
-import CTA from "components/call-to-action";
 import { CategoryData } from "components/CategoryData";
+import { CTA } from "components/CTA";
 import { Header } from "components/Header";
 import Post from "components/post";
 import { isValidDateString } from "components/PypDate/utils";
