+++ pages/[city]/index.tsx
@@ -1,8 +1,8 @@
 import cities, { ICategoryData } from "@mauriciorobayo/pyptron";
 import { TheMoneytizer } from "components/Ads";
 import { Aside } from "components/Aside";
-import CTA from "components/call-to-action";
 import { CityData } from "components/CityData";
+import { CTA } from "components/CTA";
 import { Header } from "components/Header";
 import { getLocalLongDateString } from "components/PypDate/utils";
 import { AMERICA_BOGOTA, CityType, dateParts, isCity } from "lib/utils";
